The Inting

+++ src/includes/animation.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "intalias.hpp"
 #include "spritesheet.hpp"
 #include "time.hpp"
 #include <vector>
@@ -6,14 +7,14 @@
 struct AnimationData {
     SpriteSheet Sheet;
     std::vector<SpriteSheet::Coordinates> FrameCoords;
-    size_t FrameIndex = 0;
+    u64 FrameIndex = 0;
     Image GetFrame();
     void AdvanceFrame();
 };
 
 struct Animator {
     AnimationData Data;
-    size_t MSWait;
+    u64 MSWait;
     Timer::Instance Timer;
 public:
     void Update();
